FBOT/app/requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.1
openai>=1.40.0
aiohttp>=3.8.0
aiohttp-cors>=0.7.0
pyngrok>=7.0.0
aiomysql>=0.2.0
pymysql>=1.1.0
